# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "fluentlo API",
        "description": "Laravel 12 + Sanctum + Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/app/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "operationId": "appAuthRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "shahab"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/app/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "operationId": "appAuthLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/v1/app/auth/{provider}/redirect": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Redirect to social provider",
                "operationId": "appAuthSocialRedirect",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "facebook"
                            ]
                        },
                        "example": "google"
                    },
                    {
                        "name": "test",
                        "in": "query",
                        "description": "Return JSON instead of 302 (diagnostic)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to provider"
                    },
                    "422": {
                        "description": "Unsupported provider"
                    }
                }
            }
        },
        "/v1/app/auth/{provider}/callback": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Handle social provider callback",
                "operationId": "appAuthSocialCallback",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "facebook"
                            ]
                        },
                        "example": "google"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Authorization code from provider",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Opaque state (ignored in stateless)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "error",
                        "in": "query",
                        "description": "Provider error (if user denied)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Social authentication successful"
                    },
                    "422": {
                        "description": "Social authentication failed"
                    }
                }
            }
        },
        "/v1/app/auth/validate-token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Validate Sanctum token",
                "operationId": "appAuthValidateToken",
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Token is invalid"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/app/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout user and revoke token",
                "operationId": "appAuthLogout",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/app/auth/logout-all": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout from all devices",
                "operationId": "appAuthLogoutAll",
                "responses": {
                    "200": {
                        "description": "Logout from all devices successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}